Reorganize directory structure of end-to-end tests (#2708)
* Reorganize directory structure of end-to-end tests

* change path
