+++ tests_e2e/scenarios/scripts/collect_logs.sh
