+++ tests_e2e/scenarios/tests/bvts/custom_script.py
@@ -3,7 +3,7 @@ import os
 import uuid
 import sys
 
-from tests_e2e.scenario_utils.extensions.CustomScriptExtension import CustomScriptExtension
+from tests_e2e.scenarios.modules.CustomScriptExtension import CustomScriptExtension
 
 
 def main(subscription_id, resource_group_name, vm_name):
