+++ tests_e2e/scenarios/tests/check_agent_version.py
